EigenMicrostates

Documentation for EigenMicrostates.

References

Hu et al (2019)

Sun et al (2021)

EigenMicrostates.EigenEnsemble

EigenMicrostates.EigenMicrostate

EigenMicrostates.amplitudes_correlation

EigenMicrostates.amplitudes_ensemble

EigenMicrostates.order_param_correlation

EigenMicrostates.reshape_as_matrix

EigenMicrostates.weights_correlation

EigenMicrostates.weights_ensemble

EigenMicrostates.EigenEnsemble — Type

EigenEnsemble(A)

Decompose eigen ensembles (eigen microstates and their evolution) from an statistical ensemble array A.

The last dimension of A should be the time dimension. The statistical ensemble array A does not need to be normalized.

See https://doi.org/10.1088/1572-9494/abf127 for more details.

If S::EigenEnsemble represents eigen ensmebles of a N-dimensional system (e.g. a canonical ensmeble of an N-dimensional Ising model), then:

S.U is a N+1-dimensional array, S.U[:, ..., :, i] is the i-th eigen microstate;

S.sigma is a vector, representing propability amplitudes of eigen microstates, sum(S.sigma.^2) == 1.0;

S.V is a matrix, each column representing the dynamic evolution of the corresponding eigen microstate.

Eigen ensembles are arranged in descending order of propability amplitudes.

Use ndims(S) to get the dimension of the system under study (N).

EigenMicrostates.EigenMicrostate — Type

EigenMicrostate(C, system_size)

Eigen microstates and weights of correlation matrix C.

system_size is the size of the system under study, used to reshape the eigen microstates. C is transfered to a symmetric matrix first, in which only upper triangular part of C is used.

See https://doi.org/10.1007/s11433-018-9353-x for more details.

If S::EigenMicrostate represents eigen microstates of a N-dimensional system (e.g. a canonical ensmeble of an N-dimensional Ising model), then:

S.U is a N-dimensional array, S.U[:, ..., :, i] is the i-th eigen microstate;

S.weights is a vector, representing propability of eigen microstates, sum(S.weights) == 1.0.

Eigen microstates are arranged in descending order of propability.

Use ndims(S) to get the dimension of the system under study (N).

EigenMicrostates.amplitudes_correlation — Method

amplitudes_correlation(C::AbstractMatrix) -> amplitudes

Probability amplitudes of eigen microstates via decomposing correlation matrix C.

EigenMicrostates.amplitudes_ensemble — Method

amplitudes_ensemble(A) -> sigma

Probability amplitude vector sigma of eigen ensembles of ensemble array A.

A does not need to be normalized.

EigenMicrostates.order_param_correlation — Method

order_param_correlation(C::AbstractMatrix) -> order_param

Order parameter as maximum eigenvalue of correlation matrix C.

EigenMicrostates.reshape_as_matrix — Method

reshape_as_matrix(A)

Reshape an array A of dimension N (N>1) into a matrix, by collapsing all the dimensions except the last one into a single dimension.

EigenMicrostates.weights_correlation — Method

weights_correlation(C::AbstractMatrix) -> weights

Probability ("weights") of eigen microstates via decomposing correlation matrix C.

EigenMicrostates.weights_ensemble — Method

weights_ensemble(A) -> weights

Probability ("weights") of eigen ensembles of ensmeble array A. weights = sigma.^2.

A does not need to be normalized.